UserSchema, PreSave hook, checkPassword Completed
userModels.js --- Filled out the UserSchema to require a unique lowercase string username and require a string password.

PreSave hook that hashes the password and sets the user password to the hash.

checkPassword function that compares the password to the hashed password stored on the user and returns true to the callback if the match.